% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{env.params}
\alias{env.params}
\title{Exposure-related parameter settings}
\format{
A data frame with 4 rows and 12 variables
}
\usage{
env.params
}
\description{
A dataset containing exposure-related parameters.
}
\keyword{datasets}
