% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geno.params}
\alias{geno.params}
\title{Genotype-related parameter settings}
\format{
A data frame with 4 rows and 9 variables
}
\usage{
geno.params
}
\description{
A dataset containing genotype-related parameters.
}
\keyword{datasets}
