% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.critical.results.GxE.R
\name{get.critical.results.GxE}
\alias{get.critical.results.GxE}
\title{Summarizes the main results}
\usage{
get.critical.results.GxE(
  scenario = NULL,
  pheno.model = NULL,
  geno.model = NULL,
  env.model = NULL,
  empirical.power = NULL
)
}
\arguments{
\item{scenario}{Scenario number}

\item{pheno.model}{Type of the outcome; 0 for binary and 2 for continuous}

\item{geno.model}{Genetic model; 0 for binary and 1 for additive}

\item{env.model}{Model of the enviromental explosure}

\item{empirical.power}{Estimated empirical power}

\item{sample.sizes.required}{Number of cases and controls or number of subjects required to achieve the desired power}
}
\value{
A table containing the following items:
\code{genetic.model} Model of the genetic determinant
\code{environment.model} Model of the environmental determinant
\code{number.of.subjects.required} Number of subjects required to achieve the desired power under a quantitative outcome model
\code{empirical.power} Estimated empirical power under each model
}
\description{
Gets the number of cases and controls or subjects and the empirical and theoretical power under each model and prints a summary on the screen
}
\author{
Gaye A.; Westerman K.
}
\keyword{internal}
