% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.analysis.GxE.R
\name{glm.analysis.GxE}
\alias{glm.analysis.GxE}
\title{Carries out regression analysis}
\usage{
glm.analysis.GxE(pheno.model = NULL, observed.data = NULL, g.idx = NULL)
}
\arguments{
\item{pheno.model}{Type of outcome; 0=binary and 1=continuous}

\item{observed.data}{A dataframe that contains covariates and outcome data}
}
\value{
A vector containing the beta, standard-error, z-statistic, and p-value for each of the covariates
}
\description{
Fits a conventional unconditional logistic regression model with a binary or continuous phentype as outcome and the genetic, environmental, interaction determinants as covariates.
}
\author{
Gaye A.; Westerman K.
}
\keyword{internal}
