% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno.params}
\alias{pheno.params}
\title{Phenotype-related parameter settings}
\format{
A data frame with 4 rows and 8 variables
}
\usage{
pheno.params
}
\description{
A dataset containing phenotype-related parameters.
}
\keyword{datasets}
