% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.espresso.GxE.RV.R
\name{run.espresso.GxE.RV}
\alias{run.espresso.GxE.RV}
\title{Runs a full ESPRESSO.GxE.RV analysis}
\usage{
run.espresso.GxE.RV(
  simulation.params = NULL,
  pheno.params = NULL,
  geno.params = NULL,
  env.params = NULL,
  scenarios2run = 1
)
}
\arguments{
\item{simulation.params}{general parameters for the scenario(s) to analyse}

\item{pheno.params}{paramaters for the outcome variables}

\item{geno.params}{parameters for the genetic determinant(s)}

\item{env.params}{parameters for the environmental determinant}

\item{scenarios2run}{the indices of the scenarios one wish to analyse}
}
\value{
a summary table that contains both the input parameters and
the results of the analysis
}
\description{
This function calls the functions required to run a full ESPRESSO.GxE.RV analysis
where the model consists of an outcome (binary or continuous) determined by two interacting
covariates (a SNP  and an environmental exposure)
}
\examples{
{
  
# load the table that hold the input parameters; each of the table
# hold parameters for 4 scenarios:
# scenario 1: a binary outcome determined by a binary SNP and binary exposure 
# and an interaction between the the genetic variant and the environmental exposure
# scenario 2: a binary outcome determined by an additive SNP and continuous 
# exposure and an interaction between the the genetic variant and the environmental exposure
# scenario 3: a quantitative outcome determined by a binary SNP and binary exposure, 
# and an interaction between the the genetic variant and the environmental exposure
# scenario 4: a quantitative outcome determined by an additive SNP and continuous 
# exposure and an interaction between the the genetic variant and the environmental exposure

# run the function for the first two scenarios, two binomial models
run.espresso.GxE.RV(simulation.params, pheno.params, geno.params, env.params, scenarios2run=c(1,2))

}

}
\author{
Gaye A.; Westerman K.
}
