% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.pheno.qtl.GxE.R
\name{sim.pheno.qtl.GxE}
\alias{sim.pheno.qtl.GxE}
\title{Simulates continuous outcome data}
\usage{
sim.pheno.qtl.GxE(
  numsubjects = NULL,
  pheno.mean = NULL,
  pheno.sd = NULL,
  genotype = NULL,
  geno.efkt = NULL,
  geno.frac_causal = NULL,
  environment = NULL,
  env.efkt = NULL,
  interaction = NULL,
  int.efkt = NULL
)
}
\arguments{
\item{pheno.mean}{statistical mean}

\item{pheno.sd}{standard deviation}

\item{genotype}{Genotype matrix (N x g.nvar)}

\item{geno.efkt}{Effect of the genetic variant}

\item{geno.frac_causal}{Fraction of causal genetic variants}

\item{environment}{Exposure data for environment}

\item{env.efkt}{Effect of the environmental determinants}

\item{interaction}{Interaction matrix (N x g.nvar)}

\item{int.efkt}{Interaction effect}

\item{num.subjects}{Number of subjects to simulate}
}
\value{
A dataframe of phenotype
}
\description{
Uses the effects data of the determinants to construct a linear predictor(LP). The outcome is normally distributed variable generated with a mean equal to LP and a standard deviation of 1. Some error is then added to the simulated outcome to obtained the observed outcome.
}
\author{
Gaye A.; Westerman K.
}
\keyword{internal}
