% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation.params}
\alias{simulation.params}
\title{Simulation parameter settings}
\format{
A data frame with 4 rows and 11 variables
}
\usage{
simulation.params
}
\description{
A dataset containing simulation parameters.
}
\keyword{datasets}
